Separate client/service into multiple files
