+++ plusdeck/dbus/client/__main__.py
@@ -1,0 +1,3 @@
+from plusdeck.dbus.client.cli import main
+
+main()
