+++ plusdeck/dbus/service/__main__.py
@@ -1,0 +1,3 @@
+from plusdeck.dbus.service.cli import main
+
+main()
