+++ plusdeck/dbus/service/cli.py
@@ -4,52 +4,17 @@ import os
 from typing import Optional
 
 import click
-from sdbus import (  # pyright: ignore [reportMissingModuleSource]
-    request_default_bus_name_async,
-)
 
 from plusdeck.cli import LogLevel
 from plusdeck.config import GLOBAL_FILE
-from plusdeck.dbus.error import handle_dbus_error
-from plusdeck.dbus.interface import DBUS_NAME, DbusInterface, load_client
 from plusdeck.dbus.select import (
     select_default_bus,
     select_session_bus,
     select_system_bus,
 )
+from plusdeck.dbus.service import serve
 
 logger = logging.getLogger(__name__)
-
-
-async def service(config_file: Optional[str] = None) -> DbusInterface:
-    """
-    Create a configure DBus service with a supplied config file.
-    """
-
-    client = await load_client(config_file)
-    iface = DbusInterface(client, config_file=config_file)
-
-    logger.debug(f"Requesting bus name {DBUS_NAME}...")
-    await request_default_bus_name_async(DBUS_NAME)
-
-    logger.debug("Exporting interface to path /...")
-
-    iface.export_to_dbus("/")
-
-    logger.info(f"Listening on {DBUS_NAME} /")
-
-    return iface
-
-
-async def serve(config_file: Optional[str] = None) -> None:
-    """
-    Create and serve configure DBus service with a supplied config file.
-    """
-
-    async with handle_dbus_error(logger):
-        srv = await service(config_file)
-
-        await srv.closed
 
 
 @click.command
